# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Job {
  id: string;
  jobNumber: string;
  serviceType: string;
  address: string;
  /** @nullable */
  scheduledDate?: string | null;
  /** @nullable */
  scheduledTime?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerId?: string | null;
  /** @nullable */
  haulerId?: string | null;
  /** @nullable */
  priceCents?: number | null;
  status: string;
  createdAt: string;
  updatedAt?: string;
}

export interface JobInput {
  serviceType: string;
  address: string;
  scheduledDate?: string;
  scheduledTime?: string;
  description?: string;
  customerName?: string;
  customerEmail?: string;
  customerPhone?: string;
  priceCents?: number;
}

export interface JobUpdate {
  status?: string;
  /** @nullable */
  haulerId?: string | null;
  scheduledDate?: string;
  scheduledTime?: string;
  description?: string;
  priceCents?: number;
}

export interface JobStats {
  total: number;
  active: number;
  unassigned: number;
  completed: number;
  todayRevenueCents: number;
}

export interface Hauler {
  id: string;
  userId: string;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  vehicleType?: string | null;
  /** @nullable */
  vehiclePlate?: string | null;
  serviceAreas?: string[];
  backgroundCheckConsent?: boolean;
  /** @nullable */
  backgroundCheckDate?: string | null;
  trainingCompleted?: boolean;
  /** @nullable */
  trainingCompletedDate?: string | null;
  status: string;
  /** @nullable */
  profileName?: string | null;
  /** @nullable */
  profileEmail?: string | null;
  createdAt: string;
}

export type HaulerInputDocumentsItem = {
  type?: string;
  uploaded?: boolean;
};

export interface HaulerInput {
  userId: string;
  businessName?: string;
  licenseNumber?: string;
  vehicleType: string;
  vehiclePlate: string;
  serviceAreas: string[];
  backgroundCheckConsent?: boolean;
  trainingCompleted?: boolean;
  documents?: HaulerInputDocumentsItem[];
}

export interface HaulerUpdate {
  status?: string;
  businessName?: string;
  vehicleType?: string;
  vehiclePlate?: string;
}

export interface UserProfile {
  id: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
  role: string;
}

export type ListJobsParams = {
  status?: string;
  haulerId?: string;
};

export type ListHaulersParams = {
  status?: string;
};