# lib/api-zod/src/generated/types/hauler.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Hauler {
  id: string;
  userId: string;
  /** @nullable */
  businessName?: string | null;
  /** @nullable */
  licenseNumber?: string | null;
  /** @nullable */
  vehicleType?: string | null;
  /** @nullable */
  vehiclePlate?: string | null;
  serviceAreas?: string[];
  backgroundCheckConsent?: boolean;
  /** @nullable */
  backgroundCheckDate?: string | null;
  trainingCompleted?: boolean;
  /** @nullable */
  trainingCompletedDate?: string | null;
  status: string;
  /** @nullable */
  profileName?: string | null;
  /** @nullable */
  profileEmail?: string | null;
  createdAt: string;
}