# lib/api-zod/src/generated/types/haulerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */
import type { HaulerInputDocumentsItem } from "./haulerInputDocumentsItem";

export interface HaulerInput {
  userId: string;
  businessName?: string;
  licenseNumber?: string;
  vehicleType: string;
  vehiclePlate: string;
  serviceAreas: string[];
  backgroundCheckConsent?: boolean;
  trainingCompleted?: boolean;
  documents?: HaulerInputDocumentsItem[];
}