# lib/api-zod/src/generated/types/haulerUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HaulerUpdate {
  status?: string;
  businessName?: string;
  vehicleType?: string;
  vehiclePlate?: string;
}