# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: string;
  jobNumber: string;
  serviceType: string;
  address: string;
  /** @nullable */
  scheduledDate?: string | null;
  /** @nullable */
  scheduledTime?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerId?: string | null;
  /** @nullable */
  haulerId?: string | null;
  /** @nullable */
  priceCents?: number | null;
  status: string;
  createdAt: string;
  updatedAt?: string;
}