# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  serviceType: string;
  address: string;
  scheduledDate?: string;
  scheduledTime?: string;
  description?: string;
  customerName?: string;
  customerEmail?: string;
  customerPhone?: string;
  priceCents?: number;
}