# lib/api-zod/src/generated/types/jobStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobStats {
  total: number;
  active: number;
  unassigned: number;
  completed: number;
  todayRevenueCents: number;
}