# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobUpdate {
  status?: string;
  /** @nullable */
  haulerId?: string | null;
  scheduledDate?: string;
  scheduledTime?: string;
  description?: string;
  priceCents?: number;
}