# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * KURBR API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  /** @nullable */
  fullName?: string | null;
  /** @nullable */
  email?: string | null;
  role: string;
}